refactor conversion code
as DataFrames will soon depends on Nulls and CategoricalArrays other than DataArrays
