+++ REQUIRE
@@ -1,8 +1,8 @@
 julia 0.6
 DataStructures 0.5.0
-DataFrames 0.9
+DataFrames 0.10
+CategoricalArrays 0.2.0
 NullableArrays 0.1.0
-CategoricalArrays 0.1.0
 AxisArrays 0.0.6
 NamedArrays 0.5.3
 Compat 0.34.0
