+++ src/language.jl
@@ -1,3 +1,22 @@
+# formula
+function rlang_formula(e::Expr)
+    e.head == :call || error("invalid formula object")
+    op = e.args[1]
+    if op == :&
+        op = :(:)
+    end
+    if length(e.args) > 3 && op in (:+,:*,:(:))
+        rlang_p(op,
+                rlang_formula(Expr(e.head,e.args[1:end-1]...)),
+                rlang_formula(e.args[end]))
+    else
+        rlang_p(op,map(rlang_formula,e.args[2:end])...)
+    end
+end
+rlang_formula(e::Symbol) = e
+rlang_formula(n::Number) = n
+
+
 "Create a function call from a list of arguments"
 function rlang_p(f, args...; kwargs...)
     argn = length(args)+length(kwargs)
