+++ test/convert/base.jl
@@ -294,6 +294,5 @@ b = RObject(true)
 #RCall.rlang_formula(parse("a+b"))
 @test RCall.rlang_formula(:a) == :a
 
-
 # issue 195
 @test rcopy(R"list(a=NULL)")[:a] == nothing
